+++ app/assets/stylesheets/components/_dashboard-card.scss
@@ -112,9 +112,6 @@
   .my-places-cards {
     grid-template-columns: 1fr 1fr;
   }
-  #map {
-    display: none;
-  }
 }
 
 .my-places-card-actions {
