+++ app/javascript/plugins/chat-button.js
@@ -1,0 +1,12 @@
+const display = () => {
+  const button = document.getElementById("chat-button");
+  console.log(button);
+
+  const chatDiv = document.getElementById("chatroom");
+
+  button.addEventListener('click', (event) => {
+    chatDiv.classList.toggle("chatroom-display");
+  })
+};
+
+export { display };
