+++ app/javascript/plugins/flatpickr.js
@@ -1,6 +1,4 @@
 import flatpickr from "flatpickr"
-// import "flatpickr/dist/flatpickr.min.css" // Note this is important!
-// import "flatpickr/dist/themes/airbnb.css"
 
 flatpickr(".datepicker", {});
