+++ app/views/places/index.html.erb
@@ -25,7 +25,7 @@
     <% end %>
   </div>
   <div id="map"
-        style="width: 400px;
+        style="width: 100%;
        height: 100vh;
        position: sticky;
        top: 0;"
